Fix daysUntilNewMoon and daysUntilFullMoon  + move files around, update tests
Add ExactMoon, add new initializers for it, separate out Moon vs ExactMoon
